GEOMESA-2771 Add prompt for delete-features command

+++ geomesa-tools/src/main/scala/org/locationtech/geomesa/tools/data/DeleteFeaturesCommand.scala
@@ -10,8 +10,10 @@ package org.locationtech.geomesa.tools.data
 
 import org.geotools.data.DataStore
 import org.geotools.data.simple.SimpleFeatureStore
+import org.geotools.filter.text.ecql.ECQL
 import org.locationtech.geomesa.tools._
 import org.locationtech.geomesa.tools.data.DeleteFeaturesCommand.DeleteFeaturesParams
+import org.locationtech.geomesa.tools.utils.Prompt
 import org.opengis.filter.Filter
 
 trait DeleteFeaturesCommand[DS <: DataStore] extends DataStoreCommand[DS] {
@@ -22,18 +24,21 @@ trait DeleteFeaturesCommand[DS <: DataStore] extends DataStoreCommand[DS] {
   override def execute(): Unit = {
     val sftName = params.featureName
     val filter = Option(params.cqlFilter).getOrElse(Filter.INCLUDE)
-    Command.user.info(s"Deleting features from $sftName with filter $filter. This may take a few moments...")
-    withDataStore { ds =>
-      ds.getFeatureSource(sftName) match {
-        case fs: SimpleFeatureStore => fs.removeFeatures(filter)
-        case fs => throw new IllegalStateException(s"Expected SimpleFeatureStore, got ${Option(fs).map(_.getClass.getName).orNull}")
+    val msg = s"Deleting features from schema '$sftName' with filter: ${ECQL.toCQL(filter)}"
+    if (params.force || Prompt.confirm(s"$msg\nContinue (y/n)? ")) {
+      Command.user.info(s"Deleting features, please wait...")
+      withDataStore { ds =>
+        ds.getFeatureSource(sftName) match {
+          case fs: SimpleFeatureStore => fs.removeFeatures(filter)
+          case fs => throw new IllegalStateException(s"Expected SimpleFeatureStore, got ${Option(fs).map(_.getClass.getName).orNull}")
+        }
       }
+      Command.user.info("Features deleted")
     }
-    Command.user.info("Features deleted")
   }
 }
 
 object DeleteFeaturesCommand {
   // @Parameters(commandDescription = "Delete features from a table in GeoMesa. Does not delete any tables or schema information.")
-  trait DeleteFeaturesParams extends RequiredTypeNameParam with OptionalCqlFilterParam
+  trait DeleteFeaturesParams extends RequiredTypeNameParam with OptionalCqlFilterParam with OptionalForceParam
 }
